Fix threading bug
Replace monitoring of gpg_process in separate thread to monitoring
gpg_process with asynchronous function called with GLib.Timeout.

Monitoring in separate thread caused the finished() signal of the
progress indicator widget to be emitted by that thread.
This signal caused the callback of the signal to be executed
in that same thread which in turn included the update of the
gtk progressbar widget.
Since Gtk is not thread-safe this caused weird errors every 10th
to 15th encryption / decryption run.
Errors include:
 - Xcb error:
     [xcb] Unknown sequence number while processing queue
     [xcb] Most likely this is a multi-threaded client and
           XInitThreads has not been called
     [xcb] Aborting, sorry about that.
 - Segfault in gtk_button_size_allocate() -> gtk_css_gadget_allocate()
 - Segfaults in different places

All of these problems are resolved with this commit