Rewrite python script in bash
Rewrite python script that lists all source files as relative
paths in bash.
Since this is a linux desktop application it is probably safe
to assume that a bash shell interpreter is available.
The script is not posix compatible since local variables, arrays
and string replacement are all bash features.