Automatically create file holding version string
git version script automatically writes a file containing the
current version string on each execution.
If script gets executed in a project, that does not contain
metadata from .git directory it takes the content of the created
file as version string.

This is useful for building software from source tarball where
.git content will not be available.