Add script to create a source tarball
Add script that creates a source tarball of the current state of the project.
Content matched by .gitignore is excluded but modified and non tracked files
are included.

This is the proper way to create a tarball of this project.
Tarball created by `git archive` will not include the version tag file
(which is not tracked by git) and will thus not be usable for compiling.