First commit of package build scripts
Add scripts to automatically build Ubuntu 20.04 .deb packages using
Docker containers and a lot of bash scripts.
This system is supposed to be extensible for other .deb based
Distributions as well as non .deb based Distributions.
Scripts should be as generic as possible and almost always be placed
inside the "generic" directory.
Every package target has to have a directory inside "packages"
containing the files "Dockerfile", "package.sh", "package-name.sh".
- Dockerfile should describe all dependencies needed for building
  the package.
- package.sh should create the package for the corresponding package
  target. It should be necessary to run this script inside the docker
  container defined by the corresponding Dockerfile or on a real
  system with all necessary dependencies and tools installed.
  The resulting package should be put inside the "build" directory
- package-name.sh should output the name of the resulting package file
  inside the "build" directory

All three files should in most cases be symlinks to "generic" directory.
Depending on those three script more helper scripts (that may also be
symlinks) may be necessary.

To create the package for a package target using docker use the
"create-package.sh" script and pass the name of the package target
(name of the folder) as only argument.