Replace build GitHub workflow with CI workflow
CI workflow consists of a compile and a build-package job.
The compile job tries to compile the application with multiple project
configurations.
The build-package job uses the existing shell scripts to create native
.deb and .rpm packages.

In order not to have to rebuild the docker images on every run, use
GitHubs cache action in combination with `docker image save`,
`docker image load`.
Unfortunately the cache action always saves the cache if there was a
cache miss and never saves the cache if there was a cache hit.
See:
https://github.com/actions/cache/issues/109
https://github.com/actions/cache/issues/135
https://github.com/actions/cache/issues/171

To work around this problem we save the cache with key:
`docker-cache-${{ matrix.docker_image }}-${{ github.run_id }}`
but restore with restore-key (acts as wildcard):
`docker-cache-${{ matrix.docker_image }}-`
Because every workflow run creates a new cache entry this leads to
extreme cache usage and premature eviction.