Replace subclassed Gtk.AboutDialog with function
Replace subclassed `Gtk.AboutDialog` with generator function because
subclassing `Gtk.AboutDialog` will not be possible in Gtk 4.

Trying to subclass `Gtk.AboutDialog` will fail with nasty c compiler
error because valac does not yet know about final / sealed classes.
This will hopefully be fixed soon, see:
https://gitlab.gnome.org/GNOME/vala/-/issues/1036
https://gitlab.gnome.org/GNOME/glib/-/issues/2321
https://gitlab.gnome.org/GNOME/glib/-/merge_requests/1937
https://gitlab.gnome.org/GNOME/gobject-introspection/-/merge_requests/257

This commit is part of a series of commits to prepare for porting
to Gtk 4.