Manually set window to minimum size
In some cases (Gtk 4) initial window size is calculated differently.
In order to have builds using Gtk 3 and Gtk 4 have a consistent
window size we manually do Gtk.Window.set_default_size(10, 10) which
shrinks the window to the minimum.

This commit is part of a series of commits to prepare for porting
to Gtk 4.