Remove call to Gtk.Window.set_position()
Remove call to `Gtk.Window.set_position()` because this function will
not be available in Gtk 4 because Gtk 4 is modeled after Waylands
structure.
This function also does not work under Gtk 3 + Wayland so it is probably
best to remove it entirely and let the window manager do its job.

This commit is part of a series of commits to prepare for porting
to Gtk 4.