Replace Container.add() with Grid.attach()
Replace Gtk.Container.add() with Gtk.Grid.attach()
because Gtk.Container will not be available in Gtk 4.

Because with Gtk.Grid.attach() we always need to specify row and column
to add to this requires keeping track of the current row. We do this
with a simple counter integer.
We could also add widgets below other widgets with
Gtk.Grid.attach_next_to() but I don't think this would work well with
future changes in the layout. The build_gui() function is complicated
enough and does not need extra dependencies between code sections.

This commit is part of a series of commits to prepare for porting
to Gtk 4.