Don't pass WindowType to Gtk.Window
Don't pass WindowType to constructor of Gtk.Window because passing
Gtk.WindowType will not be available in Gtk 4 because it is modeled
after Waylands structure and those window types only exist on X11.

On Gtk 3 Gtk.WindowType.TOPLEVEL is the default so not passing
WindowType does not change anything.

This commit is part of a series of commits to prepare for porting
to Gtk 4.