Manually assemble action namespace for GLib.Menu
Manually assemble the global menu (the one spawned from global menu /
burger button) because Gtk 4 does not have the
`Popover.bind_model(menu, action_namespace)` function which sets the
action namespace for the action of each menu item automatically.

This commit is part of a series of commits to prepare for porting
to Gtk 4.