Introduce GTK4 flag (not yet working)
Introduce Meson option GPG_GUI_GTK_VERSION_MAJOR that
when set to 4 will try to link with Gtk 4 instead of
(the default) Gtk 3.

Depending on whether linking with Gtk 3 or Gtk 4 a symbol
GPG_GUI_GTK_VERSION_MAJOR_THREE / GPG_GUI_GTK_VERSION_MAJOR_FOUR
will be defined which can be accessed in vala using #if / #endif.

This is only a first step in porting to Gtk 4. The code will currently
not compile with GPG_GUI_GTK_VERSION_MAJOR=4.

This commit is part of a series of commits for porting to Gtk 4.