On Gtk 4 use ToggleButton instead of RadioButton
On Gtk 4 use Gtk.ToggleButton instead of Gtk.RadioButton because
Gtk.RadioButton does not exist on Gtk 4 anymore.

This is not quite perfect. Clicking a toggled (active) Gtk.ToggleButton
will result in it being becoming untoggled (not active) and the other
Gtk.ToggleButton will get toggled (active).
This is not how Gtk.RadioButton works where clicking a toggled
Gtk.RadioButton will not un-toggle it.

This commit is part of a series of commits for porting to Gtk 4.