Add bash script to support unity builds
Meson currently does not support unity builds for vala sources out of
the box: Older versions of Meson will try to unify the resulting
C sources which fails because unifying them leads to duplicated
symbols.
Newer versions of Meson will simply refuse to unify vala sources and
print a warning message.

This commit introduces a bash script which simply checks if a unity
build is requested and replaces all vala source files with a
concatenation of those source files.

This work-around works with both the older and newer behavior of Meson
although the newer versions will still print a warning.

This is a work-around for the following Meson bug:
Vala: unity build should merge Vala code, not C code
https://github.com/mesonbuild/meson/issues/5280